Revert "Created new Screen and Tabs"
