+++ package.json
@@ -11,26 +11,16 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^13.0.0",
-    "@react-navigation/bottom-tabs": "^6.3.3",
-    "@react-navigation/material-bottom-tabs": "^6.2.3",
-    "@react-navigation/native": "^6.0.12",
-    "@react-navigation/native-stack": "^6.8.0",
     "expo": "~45.0.0",
     "expo-status-bar": "~1.3.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-native": "0.68.2",
     "react-native-gesture-handler": "^2.6.0",
-    "react-native-paper": "^4.12.4",
-    "react-native-safe-area-context": "4.2.4",
-    "react-native-screens": "~3.11.1",
     "react-native-svg": "12.3.0",
     "react-native-svg-transformer": "^1.0.0",
     "react-native-svg-uri": "^1.2.3",
-    "react-native-vector-icons": "^9.2.0",
-    "react-native-web": "0.17.7",
-    "react-navigation-fluid-transitions": "^0.3.2",
-    "react-navigation-transitions": "^1.0.12"
+    "react-native-web": "0.17.7"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
