+++ src/components/FbBtnLogin/style.js
@@ -1,0 +1,27 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  button: {
+    width: 300,
+    marginTop: 80,
+    backgroundColor: "#4267B2",
+    borderRadius: 20,
+    padding: 20,
+    flexDirection: "row",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  icon: {
+    marginRight: 12,
+  },
+  label: {
+    color: "#FFF",
+    fontSize: 20,
+    fontWeight: "700",
+    flexDirection: "row",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+});
+
+export default styles;
