+++ src/components/GgBtnLogin/style.js
@@ -1,0 +1,25 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  button: {
+    width: 300,
+    marginTop: 15,
+    backgroundColor: "#D9D9D9",
+    borderRadius: 20,
+    padding: 20,
+    flexDirection: "row",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  label: {
+    marginLeft: 12,
+    color: "#4F4949",
+    fontSize: 20,
+    fontWeight: "700",
+    flexDirection: "row",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+});
+
+export default styles;
